Moved the test files to the correct pkg too
