+++ library/src/test/java/de/rheinfabrik/heimdall2/rxjava/OAuth2AccessTokenManagerGrantNewAccessTokenTest.kt
@@ -1,12 +1,9 @@
-package de.rheinfabrik.heimdall2
+package de.rheinfabrik.heimdall2.rxjava
 
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import de.rheinfabrik.heimdall2.rxjava.grants.OAuth2Grant
-import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessToken
-import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessTokenManager
-import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessTokenStorage
 import io.reactivex.Single
 import org.junit.Test
 import java.util.*
