+++ library/src/test/java/de/rheinfabrik/heimdall2/rxjava/OAuth2AccessTokenSerializationTest.kt
@@ -1,7 +1,6 @@
-package de.rheinfabrik.heimdall2
+package de.rheinfabrik.heimdall2.rxjava
 
 import com.google.gson.Gson
-import de.rheinfabrik.heimdall2.rxjava.OAuth2AccessToken
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
